Initial work for the new logging setup for salt
